+++ qlab/problems/__init__.py
@@ -8,6 +8,7 @@ from qlab.problems import (  # noqa: F401  (import = registration)
     grover,
     maxcut,
     qft,
+    qml_classifier,
     qpe,
     shor,
     simon,
@@ -16,4 +17,4 @@ from qlab.problems import (  # noqa: F401  (import = registration)
 )
 
 __all__ = ["state_prep", "maxcut", "bernstein_vazirani", "deutsch_jozsa", "simon", "grover", "qft", "qpe",
-           "shor", "vqe"]
+           "shor", "vqe", "qml_classifier"]